Improved performance by extracting boundary box
updated frontend so that it extracts the boundary box of the digit drawn on the canvas,
adds a padding proportional to the size of the image drawn, and then sends that processed image to the backend.

This greatly improved model accuracy for numbers that were drawn smaller, or not centered.